Change JSR330Adapter to use getDeclaredConstructors
In section 4.1 of JSR 330, there is explicitly a package protected
constructor: https://jcp.org/en/jsr/detail?id=330#orig

This is also consistent with guice and others as well as errorprone:
https://errorprone.info/bugpattern/UnnecessarilyVisible
